resolve comments and move test cases
