+++ sdk/python/kfp/components/container_component_artifact_channel_test.py
@@ -14,7 +14,6 @@
 
 import unittest
 
-from kfp import dsl
 from kfp.components import component_factory
 from kfp.components import placeholders
 
@@ -36,20 +35,3 @@ class TestContainerComponentArtifactChannel(unittest.TestCase):
         self.assertRaisesRegex(AttributeError,
                                r'Cannot access artifact attribute "channel"',
                                lambda: out_channel.channel)
-
-
-class TestContainerComponentFactory(unittest.TestCase):
-
-    def test_raise_error_if_access_artifact_by_itself(self):
-
-        def comp_with_artifact_input(dataset: dsl.Input[dsl.Dataset]):
-            return dsl.ContainerSpec(
-                image='gcr.io/my-image',
-                command=['sh', 'run.sh'],
-                args=[dataset])
-
-        self.assertRaisesRegex(
-            TypeError,
-            r'Cannot access artifact by itself in the container definition.',
-            component_factory.create_container_component_from_func,
-            comp_with_artifact_input)
